Move license type to licenses package
